+++ wsac-service/WsACService/Layers/InterProcess/Abstractions/IFrameHandler.cs
@@ -1,0 +1,10 @@
+using WsACService.Layers.InterProcess.Models;
+
+namespace WsACService.Layers.InterProcess.Abstractions;
+
+public interface IFrameHandler
+{
+    public bool IsTarget(Signature signature);
+
+    public Task HandleAsync(FrameSession session, Frame frame, CancellationToken ct);
+}
