+++ wsac-service/WsACService/Layers/InterProcess/Pipe/PipeLowLevelWriter.cs
@@ -1,8 +1,8 @@
 using System.IO.Pipes;
-using WsACService.InterProcess.Abstractions;
+using WsACService.Layers.InterProcess.Abstractions;
 using WsACService.Memory;
 
-namespace WsACService.InterProcess.Pipe;
+namespace WsACService.Layers.InterProcess.Pipe;
 
 public class PipeLowLevelWriter(PipeStream pipe) : ILowLevelWriter
 {
